refactor: FortranModule を simulation/integrator へ、binding メタデータを core/material へ移動
- verification/fortran_bridge.py → simulation/integrator/fortran_module.py
  FortranModule は FortranIntegrator(simulation 層) の道具であり
  verification 専用ではないため適切な層へ移動

- verification/fortran_registry.py を分割
  - @verified_against_fortran / FortranBinding / collect_bindings
    → core/material/fortran_binding.py
    (モデル定義時に走るメタデータ収集は core の責務)
  - check_bindings のみ verification/fortran_check.py に残す

- core/material/base.py の遅延 import を廃止
  (core → verification の逆向き依存を解消)

- manforge.verification から全 symbol を引き続き re-export
  (既存の callsite は変更不要)
