+++ CLAUDE.md
@@ -134,7 +134,6 @@ autograd computes yield function gradients and the Hessian needed for the tangen
 - `simulation/driver.py`: `StrainDriver`, `StressDriver` (+ aliases `UniaxialDriver`, `GeneralDriver`) — step through strain/stress histories. Two APIs: `run(load, *, initial_stress=None, initial_state=None, collect_state)` → `DriverResult` (batch); `iter_run(load, *, initial_stress=None, initial_state=None)` → `Iterator[DriverStep]` (step-by-step, supports early break / mid-loop branching). `DriverStep` fields: `i`, `strain` (cumulative), `result` (`StressUpdateResult`), `converged`, `n_outer_iter`, `residual_inf`. `StressDriver.iter_run` accepts `raise_on_nonconverged=False` to yield non-converged steps instead of raising. `DriverResult` fields: `step_results: list[StressUpdateResult]` (primary); `stress`, `strain`, `fields` are derived properties.
 - `fitting/optimizer.py`: `fit_params()` wraps scipy.optimize (L-BFGS-B, Nelder-Mead, differential_evolution); loss defined in `fitting/objective.py`; uses drivers from `simulation/`
 - `verification/fd_check.py`: Compares AD tangent vs central finite differences
-- `verification/fortran_bridge.py`: f2py interface; calls compiled UMAT and compares output element-wise to Python (stress tol: 1e-6, tangent tol: 1e-5)
 - `verification/crosscheck_driver.py`: `CrosscheckStrainDriver` / `CrosscheckStressDriver` (both `Comparator` subclasses). Mirror `StrainDriver` / `StressDriver` respectively. `CrosscheckStressDriver.__init__` accepts `max_iter` / `tol` directly (no dict wrapper). `iter_run(load, *, initial_stress=None, initial_state=None)` yields `CrosscheckCaseResult` per step; `run(load)` returns `CrosscheckResult`. `CrosscheckCaseResult` holds `result_a`, `result_b` (raw `StressUpdateResult`) and `state_n` (step-start state) for use with `JacobianChecker.compare`. Single-step comparison: pass a 1-row `FieldHistory` with `initial_stress`/`initial_state`.
 - `verification/jacobian.py`: `JacobianBlocks` dataclass, `JacobianComparisonResult` dataclass, and `JacobianChecker(model, *, rtol=1e-8)` — class-based interface for Jacobian computation and comparison. `checker.compute(result, state_n, *, stress_trial=None)` → `JacobianBlocks` computes the residual Jacobian and exposes blocks via `jac.part[row_residual_name][col_state_name]`. `checker.compare(result_a, result_b, state_n)` → `JacobianComparisonResult` compares blocks between two results. Row names are residual labels (`effective_residual_name` from the field, or `model.dlambda_residual_name` for the Δλ slot); column names are state names (`"stress"`, `"dlambda"`, implicit non-stress keys). Default (no `residual_name` set): row and column names are identical (e.g. `jac.part["alpha"]["stress"]`). With opt-in: `Implicit(residual_name="R_alpha")` → `jac.part["R_alpha"]["stress"]`. `jac.row_names()` / `jac.col_names()` enumerate canonical name tuples; `jac.iter_blocks()` yields `("row::col", array)` pairs. `jac.full` is the full `(n_unknown, n_unknown)` matrix. `compute` accepts both `StressUpdateResult` and `ReturnMappingResult` (latter requires `stress_trial=` kwarg). Import from `manforge.verification.jacobian`.
 
@@ -154,4 +153,6 @@ State is `dict[str, anp.ndarray]` at the framework boundary (solver inputs/outpu
 
 `fortran/j2_isotropic_3d.f90` implements the same J2 algorithm as the Python reference. Subroutine names match Python class names: `j2_isotropic_3d` (core logic, f2py callable), `umat` (ABAQUS entry point), `j2_isotropic_3d_elastic_stiffness` (component check). `fortran/abaqus_stubs.f90` provides mock implementations of ABAQUS internals (SINV, SPRINC, ROTSIG) for standalone compilation. Compiled via f2py into the `j2_isotropic_3d` module. The `Dockerfile` provides a reproducible gfortran + Python 3.12 environment for Fortran builds.
 
-`FortranModule` (`verification/fortran_bridge.py`) is a thin f2py wrapper whose only job is float64 type conversion. All subroutines are called via `fortran.call(name, *args)` — the same pattern for the full routine and for sub-components.
+`FortranModule` (`simulation/integrator/fortran_module.py`) is a thin f2py wrapper whose only job is float64 type conversion. All subroutines are called via `fortran.call(name, *args)` — the same pattern for the full routine and for sub-components. Import via `from manforge.simulation.integrator import FortranModule` (also re-exported from `manforge.verification` for backward compatibility).
+
+`@verified_against_fortran` decorator, `FortranBinding`, and `collect_bindings` live in `core/material/fortran_binding.py` (importable via `from manforge.core.material import verified_against_fortran`). `check_bindings` (runtime Fortran comparison helper) lives in `verification/fortran_check.py` and is re-exported from `manforge.verification`.
