+++ src/manforge/core/material/__init__.py
@@ -11,10 +11,18 @@ from manforge.core.material.bases import (
     MaterialModelPS,
     MaterialModel1D,
 )
+from manforge.core.material.fortran_binding import (
+    FortranBinding,
+    verified_against_fortran,
+    collect_bindings,
+)
 
 __all__ = [
     "MaterialModel",
     "MaterialModel3D",
     "MaterialModelPS",
     "MaterialModel1D",
+    "FortranBinding",
+    "verified_against_fortran",
+    "collect_bindings",
 ]
