+++ src/manforge/core/material/fortran_binding.py
@@ -10,8 +10,6 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Callable
 
-import numpy as np
-
 
 @dataclass(frozen=True)
 class FortranBinding:
@@ -20,7 +18,7 @@ class FortranBinding:
     Parameters
     ----------
     subroutine:
-        Exact subroutine name as passed to :meth:`FortranModule.call`.
+        Exact subroutine name as passed to :meth:`~manforge.simulation.integrator.FortranModule.call`.
     test:
         Pytest node id of the test that verifies this binding
         (e.g. ``"tests/fortran/test_j2_bindings.py::test_check_bindings_elastic_stiffness"``).
@@ -70,7 +68,7 @@ def collect_bindings(cls) -> dict[str, FortranBinding]:
     Parameters
     ----------
     cls:
-        The class to inspect (typically a concrete :class:`MaterialModel` subclass).
+        The class to inspect (typically a concrete :class:`~manforge.core.material.MaterialModel` subclass).
 
     Returns
     -------
@@ -87,58 +85,3 @@ def collect_bindings(cls) -> dict[str, FortranBinding]:
             if hasattr(obj, "__fortran_binding__"):
                 bindings[name] = obj.__fortran_binding__
     return bindings
-
-
-def check_bindings(
-    model,
-    fortran,
-    cases: dict[str, tuple[tuple, tuple]],
-    *,
-    rtol: float = 1e-10,
-) -> dict[str, tuple[bool, float]]:
-    """Compare registered Python methods against their Fortran counterparts.
-
-    Only methods listed in *cases* are checked.  Methods that return
-    non-array types (e.g. ``dict``) are not suitable for this helper — test
-    them individually.
-
-    Parameters
-    ----------
-    model:
-        An instantiated :class:`MaterialModel`.
-    fortran:
-        A :class:`FortranModule` instance (module must be importable).
-    cases:
-        ``{method_name: (py_args, fortran_args)}``.
-        *py_args* is passed to ``getattr(model, method_name)(*py_args)``.
-        *fortran_args* is passed to ``fortran.call(subroutine, *fortran_args)``.
-    rtol:
-        Relative tolerance threshold.  A result is ``ok`` when
-        ``max_rel_err < rtol``.
-
-    Returns
-    -------
-    dict[str, tuple[bool, float]]
-        ``{method_name: (ok, max_rel_err)}``.
-
-    Raises
-    ------
-    KeyError
-        If a method in *cases* is not in ``model._fortran_bindings``.
-    """
-    bindings: dict[str, FortranBinding] = getattr(model, "_fortran_bindings", {})
-    results: dict[str, tuple[bool, float]] = {}
-
-    for method_name, (py_args, fortran_args) in cases.items():
-        binding = bindings[method_name]
-
-        py_out = getattr(model, method_name)(*py_args)
-        f_out = fortran.call(binding.subroutine, *fortran_args)
-
-        py_arr = np.asarray(py_out, dtype=float).ravel()
-        f_arr = np.asarray(f_out, dtype=float).ravel()
-
-        max_rel_err = float(np.max(np.abs(py_arr - f_arr) / (np.abs(f_arr) + 1e-14)))
-        results[method_name] = (max_rel_err < rtol, max_rel_err)
-
-    return results
